requirements.txt
Django
mysql-connector-python
pymysql
mysqlclient
psycopg2-binary
django-mysql
djangorestframework
